fix clang-tidy issue 3.3

+++ tasks/tabalaev_a_cannon_mat_mul/tests/functional/main.cpp
@@ -44,7 +44,7 @@ class TabalaevACannonMatMulFuncTests : public ppc::util::BaseRunFuncTests<InType
     input_data_ = std::make_tuple(rc, a, b);
 
     std::vector<double> c(size, 0.0);
-    LocalMatrixMultiply(a, b, c, rc);
+    LocalMatrixMultiply(a, b, c, static_cast<int>(rc));
     expected_output_ = c;
   }
